+++ docs/scripting-doc/index.d.ts
@@ -2700,7 +2700,37 @@ interface TilesetEditor {
   /**
    * Access the collision editor within the tileset editor.
    */
-  collisionEditor : TileCollisionEditor
+  readonly collisionEditor : TileCollisionEditor
+
+  /**
+   * Gets the currently selected {@link WangSet} in the "Terrain Sets" view.
+   *
+   * @since 1.9
+   */
+  readonly currentWangSet: WangSet
+
+  /**
+   * The signal emitted when {@link currentWangSet} changes.
+   *
+   * @since 1.9
+   */
+  readonly currentWangSetChanged: Signal<null>;
+
+  /**
+   * Gets the currently selected Wang color index in the "Terrain Sets" view.
+   * The value 0 is used to represent the eraser mode, and the first Wang color
+   * has index 1.
+   *
+   * @since 1.9
+   */
+  readonly currentWangColorIndex: number
+
+  /**
+   * The signal emitted when {@link currentWangColorIndex} changes.
+   *
+   * @since 1.9
+   */
+  readonly currentWangColorIndexChanged: Signal<number>;
 }
 
 /**