+++ src/tiled/mapeditor.cpp
@@ -1046,9 +1046,6 @@ void MapEditor::setCurrentBrush(EditableMap *editableMap)
 EditableWangSet *MapEditor::currentWangSet() const
 {
     auto currentWangSet = mWangDock->currentWangSet();
-    if (!currentWangSet)
-        return nullptr;
-
     return EditableManager::instance().editableWangSet(currentWangSet);
 }